requirements.txt
joblib
tqdm
plyfile
natsort
opencv-python
kornia
scipy
scikit-image
omegaconf
mmcv-full==1.7.2
numpy==1.26.4
imageio-ffmpeg==0.6.0
submodules/diff_gaussian_rasterization_df
submodules/simple-knn
thirdparty/embedding/diff-gaussian-rasterization
thirdparty/weight/gaussian_rasterization_ch3